DB-10728 Move maxExecutorCores calculation to OlapServerMaster.
