+++ hbase_sql/src/main/java/com/splicemachine/derby/lifecycle/HEngineSqlEnv.java
@@ -17,20 +17,17 @@ package com.splicemachine.derby.lifecycle;
 import java.io.ByteArrayInputStream;
 import java.io.ObjectInput;
 import java.io.ObjectInputStream;
-import java.math.BigInteger;
 import java.sql.Connection;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import java.util.stream.Stream;
 
 import com.splicemachine.SqlExceptionFactory;
 import com.splicemachine.access.HConfiguration;
-import com.splicemachine.access.SparkYarnConfiguration;
 import com.splicemachine.access.api.DatabaseVersion;
 import com.splicemachine.access.api.SConfiguration;
 import com.splicemachine.access.api.ServiceDiscovery;
+import com.splicemachine.access.configuration.HBaseConfiguration;
 import com.splicemachine.access.hbase.HBaseConnectionFactory;
 import com.splicemachine.concurrent.Clock;
 import com.splicemachine.derby.iapi.sql.PartitionLoadWatcher;
@@ -51,14 +48,10 @@ import com.splicemachine.olap.AsyncOlapNIOLayer;
 import com.splicemachine.olap.JobExecutor;
 import com.splicemachine.olap.OlapServerProvider;
 import com.splicemachine.olap.TimedOlapClient;
-import com.splicemachine.primitives.Bytes;
 import com.splicemachine.si.constants.SIConstants;
 import com.splicemachine.si.impl.driver.SIDriver;
 import com.splicemachine.uuid.Snowflake;
 import org.apache.log4j.Logger;
-
-import static com.splicemachine.access.configuration.HBaseConfiguration.SPARK_YARN_CONFIG_PATH;
-import static java.lang.System.getProperty;
 
 /**
  * @author Scott Fines
@@ -83,20 +76,20 @@ public class HEngineSqlEnv extends EngineSqlEnvironment{
     private OlapClient olapClient;
     private OperationManager operationManager;
     private ZkServiceDiscovery serviceDiscovery;
-    private static final String sparkYarnConfigZkPath =
-            HConfiguration.getConfiguration().getSpliceRootPath() + SPARK_YARN_CONFIG_PATH;
+    private static final String maxExecutorCoresZkPath =
+            HConfiguration.getConfiguration().getSpliceRootPath() + HBaseConfiguration.MAX_EXECUTOR_CORES;
 
 
-    // Find the Spark and YARN configuration from Zookeeper.
-    private static byte [] getSparkYarnConfigBytes() {
+    // Find maxExecutorCores from Zookeeper.
+    private static byte [] getMaxExecutorCoresBytes() {
         byte [] returnData = null;
         try {
-            byte [] data = ZkUtils.getData(sparkYarnConfigZkPath);
+            byte [] data = ZkUtils.getData(maxExecutorCoresZkPath);
             if (data != null && data.length > 0)
                 returnData = data;
         }
         catch (Exception | java.lang.AssertionError e) {
-            LOG.warn("Unable to find the SparkYarnConfiguration in zookeeper.");
+            LOG.warn("Unable to find maxExecutorCores in zookeeper.");
         }
         return returnData;
     }
@@ -193,240 +186,35 @@ public class HEngineSqlEnv extends EngineSqlEnvironment{
     }
 
 
-    /**
-     * Parse a Spark or Hadoop size parameter value, that may use b, k, m, g, t, p
-     * to represent bytes, kilobytes, megabytes, gigabytes, terabytes or petabytes respectively,
-     * and return back the corresponding number of bytes.  Valid suffixes can also end
-     * with a 'b' : kb, mb, gb, tb, pb.
-     * @param sizeString the parameter value string to parse
-     * @param defaultValue the default value of the parameter if an invalid
-     *                     <code>sizeString</code> was passed.
-     * @return The value in bytes of <code>sizeString</code>, or <code>defaultValue</code>
-     *         if a <code>sizeString</code> was passed that could not be parsed.
-     */
-    public static long parseSizeString(String sizeString, long defaultValue, String defaultSuffix) {
-        long retVal = defaultValue;
-        Pattern sizePattern = Pattern.compile("(^[\\d.]+)([bkmgtp]$)", Pattern.CASE_INSENSITIVE);
-        Pattern sizePattern2 = Pattern.compile("(^[\\d.]+)([kmgtp][b]$)", Pattern.CASE_INSENSITIVE);
-        sizeString = sizeString.trim();
-
-        // Add a default suffix if none is specified.
-        if (sizeString.matches("^.*\\d$"))
-            sizeString = sizeString + defaultSuffix;
-
-        Matcher matcher1 = sizePattern.matcher(sizeString);
-        Matcher matcher2 = sizePattern2.matcher(sizeString);
-        Map<String, Integer> suffixes = new HashMap<>();
-        suffixes.put("b", 0);
-        suffixes.put("k", 1);
-        suffixes.put("m", 2);
-        suffixes.put("g", 3);
-        suffixes.put("t", 4);
-        suffixes.put("p", 5);
-        suffixes.put("kb", 1);
-        suffixes.put("mb", 2);
-        suffixes.put("gb", 3);
-        suffixes.put("tb", 4);
-        suffixes.put("pb", 5);
-
-        boolean found1 = matcher1.find();
-        boolean found2 = matcher2.find();
-        Matcher matcher = found1 ? matcher1 : matcher2;
-
-        if (found1 || found2) {
-            BigInteger value;
-            String digits = matcher.group(1);
-            try {
-              value = new BigInteger(digits);
-            }
-            catch (NumberFormatException e) {
-              return defaultValue;
-            }
-            int power = suffixes.get(matcher.group(2).toLowerCase());
-            BigInteger multiplicand = BigInteger.valueOf(1024).pow(power);
-            value = value.multiply(multiplicand);
-            if (value.compareTo(BigInteger.valueOf(Long.MAX_VALUE)) > 0)
-              return Long.MAX_VALUE;
-            if (value.compareTo(BigInteger.valueOf(0)) < 0)
-              return defaultValue;
-
-            retVal = value.longValue();
-        }
-        else {
-            try {
-                retVal = Long.parseLong(sizeString);
-            }
-            catch (NumberFormatException e) {
-                return defaultValue;
-            }
-            if (retVal < 1)
-                retVal = defaultValue;
-        }
-        return retVal;
-    }
-
     @Override
     public int getMaxExecutorCores() {
         if (MAX_EXECUTOR_CORES != -1)
             return MAX_EXECUTOR_CORES;
         synchronized (HEngineSqlEnv.class) {
-            byte [] sparkYarnConfigBytes = getSparkYarnConfigBytes();
-            if (sparkYarnConfigBytes == null)
+            byte [] maxExecutorCoresBytes = getMaxExecutorCoresBytes();
+            if (maxExecutorCoresBytes == null)
                 return DEFAULT_MAX_EXECUTOR_CORES;
 
-            SparkYarnConfiguration conf = null;
+            int maxExecutorCores;
+            int numNodes;
             try {
-                ByteArrayInputStream bis = new ByteArrayInputStream(sparkYarnConfigBytes);
+                ByteArrayInputStream bis = new ByteArrayInputStream(maxExecutorCoresBytes);
                 ObjectInput in = new ObjectInputStream(bis);
-                conf = (SparkYarnConfiguration) in.readObject();
+                maxExecutorCores = in.readInt();
+                numNodes = in.readInt();
             }
             catch (Exception e) {
                 return DEFAULT_MAX_EXECUTOR_CORES;
             }
-            if (conf == null)
-                return DEFAULT_MAX_EXECUTOR_CORES;
 
-            int maxExecutorCores =
-              calculateMaxExecutorCores(conf.getYarnNodemanagerResourceMemoryMB(),
-                                        conf.getDynamicAllocationEnabled(),
-                                        conf.getExecutorInstances(),
-                                        conf.getExecutorCores(),
-                                        conf.getExecutorMemory(),
-                                        conf.getDynamicAllocationMaxExecutors(),
-                                        conf.getExecutorMemoryOverhead(),
-                                        conf.getYarnExecutorMemoryOverhead(),
-                                        conf.getNumNodes());
-            if (conf.getNumNodes() > 0) {
-                numSparkNodes = conf.getNumNodes();
+            if (numNodes > 0 && maxExecutorCores > 0) {
+                numSparkNodes = numNodes;
                 MAX_EXECUTOR_CORES = maxExecutorCores;
             }
             else
                 return DEFAULT_MAX_EXECUTOR_CORES;
         }
         return MAX_EXECUTOR_CORES;
-    }
-
-    /**
-     * Estimate the maximum number of Spark executor cores that could be running
-     * simultaneously given the current YARN and splice.spark settings.
-     *
-     * @return The maximum number of Spark executor cores.
-     * @notes See @link https://spark.apache.org/docs/latest/configuration.html
-     *        for more information on spark configuration properties.
-     */
-    public static int
-    calculateMaxExecutorCores(String memorySize,                         // yarn.nodemanager.resource.memory-mb
-                              String sparkDynamicAllocationString,       // splice.spark.dynamicAllocation.enabled
-                              String executorInstancesString,            // splice.spark.executor.instances
-                              String executorCoresString,                // splice.spark.executor.cores
-                              String sparkExecutorMemory,                // splice.spark.executor.memory
-                              String sparkDynamicAllocationMaxExecutors, // splice.spark.dynamicAllocation.maxExecutors
-                              String sparkExecutorMemoryOverhead,        // splice.spark.executor.memoryOverhead
-                              String sparkYARNExecutorMemoryOverhead,    // splice.spark.yarn.executor.memoryOverhead
-                              int numNodes)
-    {
-        if (numNodes < 1)
-            numNodes = 1;
-
-        int executorCores = 1;
-        if (executorCoresString != null) {
-            try {
-                executorCores = Integer.parseInt(executorCoresString);
-                if (executorCores < 1)
-                    executorCores = 1;
-            } catch (NumberFormatException e) {
-            }
-        }
-
-        // Initialize to defaults, then check for custom settings.
-        long memSize = 8192L*1024*1024, containerSize;
-        if (memorySize != null) {
-            try {
-                memSize = Long.parseLong(memorySize) * 1024 * 1024;
-            }
-            catch (NumberFormatException e) {
-            }
-            if (memSize <= 0)
-                memSize = 8192L*1024*1024;
-        }
-
-        boolean dynamicAllocation = sparkDynamicAllocationString != null &&
-                                    Boolean.parseBoolean(sparkDynamicAllocationString);
-        int numSparkExecutorCores = executorCores;
-
-        int numSparkExecutors = 0;
-        if (!dynamicAllocation) {
-            try {
-                numSparkExecutors = 1;
-                numSparkExecutors = Integer.parseInt(executorInstancesString);
-                if (numSparkExecutors < 1)
-                    numSparkExecutors = 1;
-                numSparkExecutorCores = numSparkExecutors * executorCores;
-
-            } catch (NumberFormatException e) {
-            }
-        }
-        else {
-            numSparkExecutors = Integer.MAX_VALUE;
-            if (sparkDynamicAllocationMaxExecutors != null) {
-                try {
-                    numSparkExecutors = Integer.parseInt(sparkDynamicAllocationMaxExecutors);
-                    if (numSparkExecutors <= 0)
-                        numSparkExecutors = Integer.MAX_VALUE;
-                }
-                catch(NumberFormatException e){
-                }
-            }
-        }
-
-        long executorMemory = 1024L * 1024 * 1024; // 1g
-        if (sparkExecutorMemory != null)
-            executorMemory =
-                parseSizeString(sparkExecutorMemory, executorMemory, "b");
-
-        long sparkMemOverhead = executorMemory / 10;
-        if (sparkExecutorMemoryOverhead != null) {
-            try {
-                sparkMemOverhead =
-                    parseSizeString(sparkExecutorMemoryOverhead, sparkMemOverhead, "m");
-            } catch (NumberFormatException e) {
-            }
-        }
-        long sparkMemOverheadLegacy = executorMemory / 10;
-        if (sparkYARNExecutorMemoryOverhead != null) {
-            try {
-                sparkMemOverheadLegacy =
-                    parseSizeString(sparkYARNExecutorMemoryOverhead, sparkMemOverheadLegacy, "m");
-            } catch (NumberFormatException e) {
-            }
-        }
-        sparkMemOverhead = Math.max(sparkMemOverheadLegacy, sparkMemOverhead);
-
-        // Total executor memory includes the memory overhead.
-        containerSize = executorMemory + sparkMemOverhead;
-
-        int maxExecutorsSupportedByYARN =
-            (memSize / containerSize) > Integer.MAX_VALUE ?
-                                        Integer.MAX_VALUE : (int)(memSize / containerSize);
-
-        if (maxExecutorsSupportedByYARN < 1)
-            maxExecutorsSupportedByYARN = 1;
-
-        maxExecutorsSupportedByYARN *= numNodes;
-
-        int maxExecutorCoresSupportedByYARN = maxExecutorsSupportedByYARN * executorCores;
-
-        if (dynamicAllocation) {
-            if (numSparkExecutors < 1)
-                numSparkExecutors = 1;
-            numSparkExecutorCores = ((long)numSparkExecutors * executorCores) > Integer.MAX_VALUE ?
-                                     Integer.MAX_VALUE : numSparkExecutors * executorCores;
-        }
-
-        if (numSparkExecutorCores > maxExecutorCoresSupportedByYARN)
-            numSparkExecutorCores = maxExecutorCoresSupportedByYARN;
-
-        return numSparkExecutorCores;
     }
 
     /**
