+++ splice_access_api/src/main/java/com/splicemachine/access/configuration/HBaseConfiguration.java
@@ -133,9 +133,9 @@ public class HBaseConfiguration implements ConfigurationDefault {
     public static final String SERVERS_PATH = "/servers";
 
     /**
-     * Path in ZooKeeper for registering the SparkYarnConfiguration
+     * Path in ZooKeeper for registering maxExecutorCores
      */
-    public static final String SPARK_YARN_CONFIG_PATH = "/sparkYarnConfig";
+    public static final String MAX_EXECUTOR_CORES = "/maxExecutorCores";
 
 
     public static final String DDL_PATH="/ddl";
